First commit
This is a basic implementation of the the four main methods in
the hashgraph protocol.

The way in which the network communicates it's not specified.
Each new `Node` will provide an implementation of `Peer` which
will handle the communication however it pleases, either by
process communication, network, etc.

The protocol used here is bincode, but I did the test with JSON
and still works as expected.