Use hash structures when it makes sense
`Hashgraph` needs to use `BTreeMap` because the order it's
important to make the serialization deterministic. So I created
the appropriate traits for `EventHash` to handle that and used
b-tree like structures every time I needed a map or a set with
`EventHash`.

However, that's not always the most efficient choice. In this
commit, I implemented the `Hash` trait for `EventHash` and used
hash based structures in the places where order isn't
important.