Create tests for node
They are not really unitary, since they depend on one implementation of
`Hashgraph`. In an ideal world, one of the mocking libraries of rust
would work correctly in our project, but after spending way too much
time trying to make them happy I got nothing but sadness and was forced
down this road.

We could also create a custom, test only `Hashgraph` implementation
therefore making this unit test again. But the line it's so blurry right
now that it doesn't really maatter. I'm willing to open a ticket
tracking this problem, though.