Remove all trace of mutability from Node
Finally, remove all `&mut self` references.

In order to do that, we have to abstract away the internal
state.

Well, we don't _have to_, in theory, we could leave that state
in the `Node` struct and be happy about that. But that's would
require one extra mutex per mutable internal state. That's way
too many mutexes. This way we only add one.

There's an strong argument for moving `network` into the
internal state structure. I decided not to do it to reduce the
amount of change that had to be done, but deep down I think it
belongs there.