Create a Dummy implementation of peer
This dummy implementation it's very simple:

It will start two threads, one for syncing and another one for
responding syncs. The two threads will communicate through each
other using an internal channel. They will communicate with a
serialized version of the message (i.e.
`(Hashgraph, EventHash)` where the first one it's the current
state of the Hashgraph and the second one it's the head).

In order to do this, I had to add the `Deserialize` trait to
`Hashgraph` and `EventHash`, which was not trivial. And in
order to that, I had to refactor a bunch of stuff so I don't
have to write custom `Deserialize` implementations because they
are ridiculously complicated.