Finish dummy implementation
This commit finishes the implementation of the dummy client and
sets up a simple binary to start X number of nodes.

It successfully runs in the sense that it starts the nodes and
starts transmitting messages.

It includes some changes too:

- Simplify the `Peer` trait to not include a function that's
really not needed.
- Make `Node` `Sync` + `Send`. In order to do that, I had to
change some of the way it works. In order to do that, I had
to delete all reference to `RefCell` and `Rc`. As a consequence
there was an small degrade of the behavior: Previously, `Node`
would be able to handle an implementation of `Hashgraph` in the
object itself and the same time accept a _different_
implementation for the remote `Hashgraph`. However, with this
change, they have to be the same. It's a very small lose and we
can rethink it in the future, though.
- Move `Network` to `NodeInternalState` finally and reduce the
number of `Mutex`es.
- Remove all notions of `Channel` in the `Node` because it's
not really necessary.
- Require all the `trait` to also be `Send` and `Sync`.
- Fix a couple edge cases, particularly with the distribution
of votes in the first round, as the maximum round was being
wrongly calculated (it was returning the number of rounds and
not the maximum id).

It doesn't successfully run in the sense that there's still a
mistake in the `Hashgraph` algorithm that makes the `Node` fail
because the event wasn't found when trying to sync remote.

However, that error isn't introduced (but discovered) with this
changes and will be fixed in a future pr with some of the
integration tests.