Fix subtraction overflow on witness search
Witness search requires looking at the the previous round.
The number of rounds is a natural number, so the id of rounds
should be a `usize`. However, that means that in witness search
for the round `0`, we might need to look in a round isn't in
the domain of the type `usize` and we will get a subtract
overflow from the runtime.

Prevent that by handling round zero as an edge case.