Fix last errors and make a small optimization
When merging the two hashgraphs, I was merging them with the
events in no order. This has an interesting property:

Suppose that the remote hashgraph says that it has three new
events: e1, e2 and e3 and in that order they appear in the
vector. Suppose also that e2 is the head and e3 is the self
parent of e2.

When we try to add e2, we discover that we don't have e3 in our
hashgrasph and mark it as invalid. Because we aren't adding the
remote head, later on,m when create our new head with the
remote head as the other parent, it will point to an event that
isn't in the hashgraph and cause errors down the line when
trying to reach for the events the parents can see in the head.

In order to prevent this, we order the events in the difference
by their height in their graph, in such a way that parents
appear always earlier in the vector while children appear last.
In our previous example, e3 would be before e2, it'd be
correctly added (either the parents of e3 are already in the
local hashgraph or the e1 is parent of e3, in which case it
will appear first) and when it arrives to the remote head, e2,
it will be marked as valid.

Also, I made a small improvement: If the difference between the
remote and local hashgraph is an empty list, not even try to
create a new head. This prevents the creation of useless events
that report no actual change in the hashgraph. This happens a
lot if you happen to sync with the same node more than once in
a row.