Create an implementation of Peer using TCP
There's a lot of duplicated code from Dummy. There's a
non-written rule in software development that says once you
repeat something more than two times, you should abstract it,
not sooner and not later. This is the second time. If the next
implementations turns out to need the same infrastructure, it's
time to abstract the common parts. For now, let's leave it as
is, as it might be coincidental and I don't want to add
unnecessary layers of complexity.

Other than that, pretty straight forward once you understand
Dummy. The only difference is that instead of communicating
between processes now it communicates through TCP streams.

The network protocol is very simple:

- A connection open to a peer is interpreted as a request to
sync.
- On connection open, the peer sends current head and hashgraph
to the client. In bincode.
- After that, connection is closed.
- No other transmission so far.

Right now, that might be sent into multiple tcp packages,
depending on the hashgraph's size. Over time, that number will
grow. An easy way to ease this would be to make the
`HashgraphWire` include just the subset of the graph that is
still being worked on, as mentioned in the paper. But this pr
doesn't address that.

This closes #4.