VM: Start parsing instructions
Set the fundations on how to parse the VM.

Change the iterator to not be peekable, as this particular binary code
doesn't need it.

Change the structure to use `failure::Error` instead of a custom `enum`.

Also, should the `TryFrom` work creating a stream? Maybe, maybe... Let's
do that later, though.

Signed-off-by: Agustin Chiappe Berrini <jnieve@gmail.com>