VM: Finish parsing binaries
All the instructions parsed up till here had been cross tested with the
original VM.

The `TryFrom<Vec<u8>>` could easily be a `From<Vec<u8>>` for a `Program`
that has `Iterable<Result<Instruction, Error>>` instead of a vector.
That would arguably be better, but that's bikeshading and that should be
enough to unblock me.

Signed-off-by: Agustin Chiappe Berrini <jnieve@gmail.com>