Fix failing test
Some tests in swirlds were timing out. The problem was that the
function that adds a new head, executed at the beginning of the
creation of a new node, requires locking of the mutex that holds
the state. It keeps that lock till the end. However, in the
process, we call `add_event`, which also requires access to that
lock. Because it is already held, it keeps waiting forever for
it to be released.

The solution is to shorten the life of the first lock by scoping
the variable in a different method.

Signed-off-by: Agustin Chiappe Berrini <jnieve@gmail.com>