VM: Very simple memory allocator
The allocator consists of three structures: A reference to the memory
used so far to get the capacity, a vector of free chunks (every chunk is
a pair with starting and ending address) sorted in descending order by
space in the chunk and a map from allocated address to size allocated.

`malloc` uses the free chunks to look for space to allocate and saves it
in the allocated spaces map. `free` gets the size of the address from
the allocated spaces map and then adjusts the free chunks structure to
remove the freed space.

The free chunks are sorted in order to help with memory fragmentation:
It helps finding the minimum chunk with enough space when allocating
(that reduces the chances of overly small chunks). Also, when freeing
space we would first look for the bigger adjacent space and if it
exists, we would just merge the two of them together.

That makes both `free` and `malloc` have complexity `O(n)`, where `n` is
the size of the `free_chunks` structure.

Signed-off-by: Agustin Chiappe Berrini <jnieve@gmail.com>