Improve unsafe usage
This comes in two parts:

- Remove it places in which is possible. For example, the
`deserialize_array` function needed to be unsafe because it also needed
to be generic. However, it's also used in only one place, so it doesn't
really _need_ to be generic. By specializing its typing, we remove
`unsafe` usage.

- Move all the unsafe to a few, well defined places instead of it being
all around.

TODO: Do we really need to call those unsafe scanf? They were necessary
to keep compatibility with the original vm, but we already diverged...
Maybe there's a better method?

Signed-off-by: Agustin Chiappe Berrini <jnieve@gmail.com>